@@ -1,25 +1,26 @@
 {
-  "name": "bpl8",
-  "version": "1.1.3",
+  "name": "better-angels-web",
+  "version": "1.0.0",
   "description": "",
   "main": "app.js",
-  "author": "Colton Brown <coltonTB@me.com> (http://coltontbrown.com/)",
+  "author": "Colton Brown <colton@coltonbrown.co> (http://coltonbrown.co/)",
   "license": "ISC",
   "config": {
-    "s3BucketName": "",
+    "s3BucketName": "better-angels-web",
     "lambdaRegion": "",
     "lambdaFunctionName": ""
   },
   "dependencies": {
     "async": "^2.5.0",
-    "aws-sdk": "^2.86.0",
+    "aws-sdk": "^2.141.0",
     "aws-serverless-express": "^2.1.3",
+    "axios": "^0.17.0",
     "babel-cli": "^6.23.0",
     "babel-core": "^6.23.1",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^6.3.2",
     "babel-plugin-inline-json-import": "^0.2.1",
-    "babel-plugin-styled-components": "^1.1.7",
+    "babel-plugin-styled-components": "^1.2.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
@@ -31,33 +32,46 @@
     "css-name": "^1.0.0",
     "dotenv": "^4.0.0",
     "dotty": "0.0.2",
-    "eslint-plugin-react": "^6.10.0",
+    "email-validator": "^1.1.1",
     "eslint": "^3.16.1",
-    "express-basic-auth": "^1.1.1",
+    "eslint-plugin-react": "^6.10.0",
     "express": "^4.14.1",
+    "express-basic-auth": "^1.1.1",
+    "found": "^0.3.4",
+    "found-scroll": "^0.1.2",
     "history": "^3.2.0",
+    "humanparser": "^1.5.0",
+    "moment": "^2.19.1",
     "prop-types": "^15.5.10",
-    "react-dom": "^15.4.2",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-dropdown": "^1.3.2",
     "react-jsonschema-form": "^0.49.0",
-    "react-router": "^3.0.2",
+    "react-redux": "^5.0.6",
     "react-svg": "^2.1.19",
-    "react": "^15.4.2",
-    "scroll-doc": "^0.1.3",
+    "redux": "^3.7.2",
+    "redux-ghost": "git+ssh://git@github.com:ForGoodMeasure/redux-ghost.git",
+    "redux-thunk": "^2.2.0",
     "scroll": "^2.0.0",
-    "styled-components": "^2.1.1",
+    "scroll-doc": "^0.1.3",
+    "serialize-javascript": "^1.4.0",
+    "stringy": "^0.1.0",
+    "stripe": "^5.2.0",
+    "styled-components": "^2.2.1",
     "styled-normalize": "^2.2.1",
-    "superagent": "^3.5.2",
+    "superagent": "^3.8.0",
+    "svg-inline-loader": "^0.8.0",
     "webpack": "^2.2.1"
   },
   "scripts": {
     "clone-content": "aws s3 cp s3://$npm_package_config_s3BucketName/content/content.json ./content/content.json",
     "clean": "rm -rf dist/* && mkdir dist/static dist/lambda dist/static-pages && cp -r src/images dist/static",
-    "build-server": "./node_modules/.bin/webpack",
+    "build-server": "webpack",
     "deploy-lambda": "serverless deploy -f hello",
     "upload-assets": "aws s3 cp --recursive ./dist/static s3://$npm_package_config_s3BucketName --acl public-read",
     "upload-content": "aws s3 cp --recursive ./content s3://$npm_package_config_s3BucketName/content --acl public-read",
     "deploy": "npm run clean && npm run build-server && npm run deploy-lambda && npm run upload-assets && npm run upload-content",
-    "local": "./node_modules/.bin/nodemon -w ./dist",
+    "local": "nodemon -w ./dist",
     "watch": "npm run build-server -- --watch"
   }
 }
