@@ -1,17 +1,21 @@
 import { getContent } from './content-manager';
 
-const getLocalContext = ({stageContext, content}) => ({
+const getLocalContext = ({stageContext, content, storeState={}, stripeKeyPub}) => ({
   assetUrl(assetPath) {
     return [stageContext.assetBase, assetPath].join('');
   },
   resourceUrl(resourcePath) {
     return [stageContext.resourceBase, resourcePath].join('')
   },
+  ghostUrl(resourcePath) {
+    return [stageContext.ghostBase, resourcePath].join('')
+  },
   getContent(namespace, key) {
     return getContent(content, namespace, key);
   },
   stageContext,
-  content
+  content,
+  storeState
 });
 
 export default getLocalContext;
