@@ -1,29 +1,30 @@
 const localStageContext = {
   stage: 'local',
   resourceBase: '',
-  assetBase: ''
+  assetBase: '',
+  gaId: '',
+  faviconUrl: ''
 };
 
-const staticSiteContext = {
-  stage: 'static',
+const prodStageContext = {
+  stage: 'prod',
   resourceBase: '',
-  assetBase: ''
+  assetBase: '',
+  gaId: '',
+  faviconUrl: ''
 };
 
 /*
   Extract stage variables from API gateway
 */
 function getStageContext(req) {
-  let stageContext;
-  const apiGatewayEventHeader = req.headers['x-apigateway-event']
-  if (apiGatewayEventHeader) {
-    try {
-      stageContext = JSON.parse(apiGatewayEventHeader).stageVariables;
-    } catch (e) {}
-  } else if (req.headers['x-static-generator']) {
-    stageContext = staticSiteContext;
+  const NODE_ENV = process.env.node_env;
+  if (NODE_ENV === 'dev') {
+    return localStageContext;
+  } else if (NODE_ENV === 'prod') {
+    return prodStageContext;
   }
-  return stageContext || localStageContext;
+  throw "Please set node_env";
 }
 
 export default getStageContext;
