@@ -1,30 +1,41 @@
+import metaContent from './meta-content';
+
 const Scaffold = (props, reactContent) =>
 `
   <!doctype html>
   <html lang="en">
-
     <head>
-      <title>Radical Machines</title>
-      <meta name="description" content="" />
+
+      ${ metaContent(props.metaContent) }
+      <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
       <meta name="author" content="" />
       <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
       <meta name="viewport" content="width=device-width, initial-scale=1">
 
       <link href="//fonts.googleapis.com/css?family=Raleway:400,300,600" rel="stylesheet" type="text/css"/>
-      <link rel="icon" type="image/png" href="http://d1tug40blkttrl.cloudfront.net/icons/dot.ico"/>
+      <link rel="icon" type="image/png" href="${ props.localContext.faviconUrl }"/>
 
       <!- StyledComponents ->
-        <style type="text/css">${props.styleSheet.getStyleTags()}</style>
+        <style type="text/css">${ props.styleSheet.getStyleTags() }</style>
       <!- StyledComponents ->
 
       <script type="text/javascript">
-        window.__locals__=${JSON.stringify({...props.localContext, imageUrls: props.imageUrls})}
+        window.__locals__=${ JSON.stringify({...props.localContext, imageUrls: props.imageUrls}) }
+      </script>
+
+      <!-- Global site tag (gtag.js) - Google Analytics -->
+      <script async src="https://www.googletagmanager.com/gtag/js?id=UA-93943838-2"></script>
+      <script>
+        window.dataLayer = window.dataLayer || [];
+        function gtag(){dataLayer.push(arguments);}
+        gtag('js', new Date());
+        gtag('config', ${ props.localContext.gaId });
       </script>
     </head>
     <body>
-      <div id="app-content">${reactContent}</div>
+      <div id="app-content">${ reactContent }</div>
       <script src="https://apis.google.com/js/api.js"></script>
-      <script type="text/javascript" src="${props.localContext.assetUrl("/assets/browser-bundle.js")}"></script>
+      <script type="text/javascript" src="${ props.localContext.resourceUrl("/assets/browser-bundle.js") }"></script>
     </body>
 
   </html>
